alter gacha to take from pool of anime & mangas

+++ srv/models/canons/canon.go
@@ -2,7 +2,7 @@ package canons
 
 type Canon struct {
 	MALType string
-	MALID float64
+	MALID int
 	Title string
 	URL string
 	ImageURL string
@@ -11,7 +11,7 @@ type Canon struct {
 func MAL2Canon(malType string, mal map[string]interface{}) Canon {
 	return Canon{ 
 		MALType: malType,
-		MALID: mal["mal_id"].(float64), 
+		MALID: int(mal["mal_id"].(float64)), 
 		Title: mal["title"].(string),  
 		URL: mal["url"].(string),  
 		ImageURL: mal["image_url"].(string)}
