+++ srv/models/waifus/waifu.go
@@ -1,7 +1,7 @@
 package waifus
 
 type Waifu struct {
-	MALID float64
+	MALID int
 	Name string
 	ImageURL string
 	URL string
@@ -9,7 +9,7 @@ type Waifu struct {
 
 func MAL2Waifu(mal map[string]interface{}) Waifu {
 	return Waifu{ 
-		MALID: mal["mal_id"].(float64), 
+		MALID: int(mal["mal_id"].(float64)), 
 		Name: mal["name"].(string),  
 		URL: mal["url"].(string),  
 		ImageURL: mal["image_url"].(string)}
